+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- TMPVisibleCharactersAnimation
 - ScaleAnimation
 
 ## [0.3.0] - 2026-07-10