+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- AnimationCurveExtension
-- TMPVisibleCharactersAnimation
-- ScaleAnimation
+- Ease Animation curves preset
+- Animation Curve Extensions
+- TextMesh Pro Visible Characters Animation component
+- Scale Animation component
 
 ## [0.3.0] - 2026-07-10
 ### Added