+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2026-07-14
 ### Added
 - Ease Animation curves preset
 - Animation Curve Extensions
@@ -42,7 +44,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/AnimationSystem/compare/0.3.0...main
-[0.2.0]: https://github.com/HyagoOliveira/AnimationSystem/tree/0.3.0/
+[Unreleased]: https://github.com/HyagoOliveira/AnimationSystem/compare/0.4.0...main
+[0.4.0]: https://github.com/HyagoOliveira/AnimationSystem/tree/0.4.0/
+[0.3.0]: https://github.com/HyagoOliveira/AnimationSystem/tree/0.3.0/
 [0.2.0]: https://github.com/HyagoOliveira/AnimationSystem/tree/0.2.0/
 [0.1.0]: https://github.com/HyagoOliveira/AnimationSystem/tree/0.1.0/